# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetRead": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "crs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrsMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "defaultVisible": {
            "default": false,
            "title": "Defaultvisible",
            "type": "boolean"
          },
          "footprint": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "MultiPolygon": "#/components/schemas/MultiPolygon",
                    "Polygon": "#/components/schemas/Polygon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPolygon"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "observedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider": {
            "$ref": "#/components/schemas/AssetProvider"
          },
          "renderConfig": {
            "$ref": "#/components/schemas/RenderConfig"
          },
          "representation": {
            "$ref": "#/components/schemas/Representation"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "siteId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siteid"
          },
          "source": {
            "discriminator": {
              "mapping": {
                "3d-tiles-url": "#/components/schemas/TilesUrlSource",
                "cesium-ion": "#/components/schemas/CesiumIonSource"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CesiumIonSource"
              },
              {
                "$ref": "#/components/schemas/TilesUrlSource"
              }
            ],
            "title": "Source"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "validFrom": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validfrom"
          },
          "validTo": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validto"
          }
        },
        "required": [
          "name",
          "representation",
          "source",
          "id",
          "siteId",
          "provider",
          "createdAt",
          "updatedAt"
        ],
        "title": "AssetRead",
        "type": "object"
      },
...
      "Body_upload_thumbnail_api_v1_sites__site_id__thumbnail_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_thumbnail_api_v1_sites__site_id__thumbnail_post",
        "type": "object"
      },
...
      "GeoPosition": {
        "additionalProperties": false,
        "properties": {
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "longitude",
          "latitude"
        ],
        "title": "GeoPosition",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SiteCreate": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetBase"
            },
            "title": "Assets",
            "type": "array"
          },
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "boundary": {
            "discriminator": {
              "mapping": {
                "MultiPolygon": "#/components/schemas/MultiPolygon",
                "Polygon": "#/components/schemas/Polygon"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Polygon"
              },
              {
                "$ref": "#/components/schemas/MultiPolygon"
              }
            ],
            "title": "Boundary"
          },
          "cameraBookmarks": {
            "items": {
              "$ref": "#/components/schemas/CameraBookmarkCreate"
            },
            "title": "Camerabookmarks",
            "type": "array"
          },
          "centroid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPosition"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to the boundary centroid when omitted."
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          }
        },
        "required": [
          "name",
          "boundary"
        ],
        "title": "SiteCreate",
        "type": "object"
      },
...
      "SiteRead": {
        "additionalProperties": false,
        "properties": {
          "areaM2": {
            "title": "Aream2",
            "type": "number"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetRead"
            },
            "title": "Assets",
            "type": "array"
          },
          "attribution": {
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "title": "Attribution",
            "type": "array"
          },
          "boundary": {
            "discriminator": {
              "mapping": {
                "MultiPolygon": "#/components/schemas/MultiPolygon",
                "Polygon": "#/components/schemas/Polygon"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Polygon"
              },
              {
                "$ref": "#/components/schemas/MultiPolygon"
              }
            ],
            "title": "Boundary"
          },
          "cameraBookmarks": {
            "items": {
              "$ref": "#/components/schemas/CameraBookmarkRead"
            },
            "title": "Camerabookmarks",
            "type": "array"
          },
          "centroid": {
            "$ref": "#/components/schemas/GeoPosition"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "name",
          "boundary",
          "centroid",
          "id",
          "slug",
          "areaM2",
          "assets",
          "cameraBookmarks",
          "createdAt",
          "updatedAt"
        ],
        "title": "SiteRead",
        "type": "object"
      },
      "SiteSummary": {
        "additionalProperties": false,
        "properties": {
          "areaM2": {
            "title": "Aream2",
            "type": "number"
          },
          "centroid": {
            "$ref": "#/components/schemas/GeoPosition"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latestObservedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestobservedat"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SiteQuality"
              },
              {
                "type": "null"
              }
            ]
          },
          "representations": {
            "items": {
              "type": "string"
            },
            "title": "Representations",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "centroid",
          "areaM2",
          "thumbnailUrl",
          "representations",
          "latestObservedAt",
          "quality",
          "createdAt",
          "updatedAt"
        ],
        "title": "SiteSummary",
        "type": "object"
      },
      "SiteUpdate": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Attribution"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "boundary": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "MultiPolygon": "#/components/schemas/MultiPolygon",
                    "Polygon": "#/components/schemas/Polygon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/MultiPolygon"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "centroid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          }
        },
        "title": "SiteUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nCatalog service for the geospatial digital twin.\n\n* **Sites** \u2014 physical places with reality captures, footprints and camera bookmarks.\n* **Assets** \u2014 derived, renderable delivery assets (3D Tiles) with provenance.\n* **Layers** \u2014 composable world layers from open data or the user's own sources.\n\nAll geometry is GeoJSON (WGS 84, RFC 7946). All timestamps are ISO 8601.\n",
    "title": "Digital Twin API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sites": {
      "get": {
        "operationId": "list_sites_api_v1_sites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteSummary"
                  },
                  "title": "Response List Sites Api V1 Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List sites",
        "tags": [
          "sites"
        ]
      },
      "post": {
        "operationId": "create_site_api_v1_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a site",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/by-slug/{slug}": {
      "get": {
        "operationId": "get_site_by_slug_api_v1_sites_by_slug__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a site by slug",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/{site_id}": {
      "delete": {
        "operationId": "delete_site_api_v1_sites__site_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a site",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "operationId": "get_site_api_v1_sites__site_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a site",
        "tags": [
          "sites"
        ]
      },
      "patch": {
        "operationId": "update_site_api_v1_sites__site_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a site",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/{site_id}/assets": {
      "get": {
        "operationId": "list_site_assets_api_v1_sites__site_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetRead"
                  },
                  "title": "Response List Site Assets Api V1 Sites  Site Id  Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a site's assets",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/{site_id}/bookmarks": {
      "get": {
        "operationId": "list_bookmarks_api_v1_sites__site_id__bookmarks_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CameraBookmarkRead"
                  },
                  "title": "Response List Bookmarks Api V1 Sites  Site Id  Bookmarks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List camera bookmarks",
        "tags": [
          "sites"
        ]
      },
      "post": {
        "operationId": "create_bookmark_api_v1_sites__site_id__bookmarks_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraBookmarkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraBookmarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save a camera bookmark",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/{site_id}/bookmarks/{bookmark_id}": {
      "delete": {
        "operationId": "delete_bookmark_api_v1_sites__site_id__bookmarks__bookmark_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bookmark Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a camera bookmark",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/sites/{site_id}/thumbnail": {
      "post": {
        "operationId": "upload_thumbnail_api_v1_sites__site_id__thumbnail_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_thumbnail_api_v1_sites__site_id__thumbnail_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a site thumbnail",
        "tags": [
          "sites"
        ]
      }
    }
  }
}